Moved Format to Utility/String
