+++ src/Application/Game.cpp
@@ -83,7 +83,7 @@
 #include "Library/Logger/Logger.h"
 
 #include "Utility/String/Ascii.h"
-#include "Utility/Format.h"
+#include "Utility/String/Format.h"
 #include "Utility/DataPath.h"
 #include "Utility/Exception.h"
