+++ src/Bin/CodeGen/CodeGenMap.h
@@ -10,7 +10,7 @@
 #include <algorithm>
 
 #include "Utility/Workaround/ToUnderlying.h"
-#include "Utility/Format.h"
+#include "Utility/String/Format.h"
 #include "Utility/String/TransparentFunctors.h"
 
 #include "CodeGenFunctions.h"
