+++ src/Library/StackTrace/StackTrace.cpp
@@ -6,7 +6,7 @@
 #   include <backward.hpp>
 #endif
 
-#include "Utility/Format.h"
+#include "Utility/String/Format.h"
 
 #ifdef __ANDROID__
