+++ src/Utility/Exception.h
@@ -2,10 +2,9 @@
 
 #include <stdexcept>
 #include <string_view>
-#include <string>
 #include <utility>
 
-#include "Format.h"
+#include "Utility/String/Format.h"
 
 class Exception : public std::runtime_error {
  public:
