+++ src/Utility/String.cpp
@@ -5,7 +5,7 @@
 #include <algorithm>
 #include <string>
 
-#include "Format.h"
+#include "Utility/String/Format.h"
 
 std::vector<char *> tokenize(char *input, const char separator) {
     std::vector<char *> retVect;
